+++ dpmModule/jobs/jobclass/adventurer.py
@@ -10,3 +10,26 @@ class MapleHero2:
     def mapleHero2(var1, var2):
         return core.BuffSkill("메이플월드 여신의 축복")
 '''
+# 모험가 법사
+# TODO : 재사용 대기시간 초기화 미적용으로 변경
+class InfinityWrapper(core.BuffSkillWrapper):
+    def __init__(self, serverlag = 3):
+        skill = core.BuffSkill("인피니티", 960, 40000, cooltime = 180 * 1000, rem = True, red = True)
+        super(InfinityWrapper, self).__init__(skill)
+        self.passedTime = 0
+        self.serverlag = serverlag
+        
+    def spend_time(self, time):
+        if self.onoff:
+            self.passedTime += time
+        super(InfinityWrapper, self).spend_time(time)
+            
+    def get_modifier(self):
+        if self.onoff:
+            return core.CharacterModifier(pdamage_indep = (70 + 4 * (self.passedTime // ((4+self.serverlag)*1000))) )
+        else:
+            return core.CharacterModifier()
+        
+    def _use(self, rem = 0, red = 0):
+        self.passedTime = 0
+        return super(InfinityWrapper, self)._use(rem = rem, red = red)
